+++ src/HeaderSort/ui/HeaderSort.scss
@@ -2,7 +2,6 @@
     .sort-header {
         display:flex;
         align-items:center;
-        border-style: solid;
         border-color: #D7D7D7;
         border-top-width: 0;
         border-right: 0;
